refactor(LobbyScreen):exported multiple components as separate modules
